Fix Anthropic max_tokens type error
Cast max_tokens to integer to fix API request error
- Anthropic API requires max_tokens as integer, not string
- env() returns string, explicit (int) cast needed
- Fixes: 'max_tokens: Input should be a valid integer' error

+++ config/services.php
@@ -42,7 +42,7 @@ return [
     'anthropic' => [
         'api_key' => env('ANTHROPIC_API_KEY'),
         'model' => env('ANTHROPIC_MODEL', 'claude-3-haiku-20240307'),
-        'max_tokens' => env('ANTHROPIC_MAX_TOKENS', 1024),
+        'max_tokens' => (int) env('ANTHROPIC_MAX_TOKENS', 1024),
         'version' => env('ANTHROPIC_VERSION', '2023-06-01'),
     ],
